+++ service/coupon-api/src/main/java/com/couponify/couponapi/application/CouponCreateService.java
@@ -3,38 +3,24 @@ package com.couponify.couponapi.application;
 import com.couponify.couponapi.presentation.request.CouponCreateRequest;
 import com.couponify.coupondomain.domain.coupon.Coupon;
 import com.couponify.coupondomain.domain.coupon.repository.CouponRepository;
-import com.couponify.coupondomain.domain.issuedCoupon.repository.IssuedCouponRepository;
-import java.time.LocalDateTime;
-import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-@Slf4j(topic = "CouponService")
+@Slf4j(topic = "CouponCreateService")
 @Service
 @Transactional(readOnly = true)
 @RequiredArgsConstructor
-public class CouponService {
+public class CouponCreateService {
 
     private final CouponRepository couponRepository;
-    private final IssuedCouponRepository issuedCouponRepository;
 
     @Transactional
     public Long create(CouponCreateRequest couponCreateRequest) {
         final Coupon coupon = CouponCreateRequest.toDomain(couponCreateRequest);
         final Coupon savedCoupon = couponRepository.save(coupon);
         return savedCoupon.getId();
-    }
-
-    @Transactional
-    public void expire() {
-        List<Coupon> expiredCoupons = couponRepository.findExpiredCoupons(LocalDateTime.now());
-        if (expiredCoupons.isEmpty()) {
-            return;
-        }
-        expiredCoupons.forEach(Coupon::expire);
-        log.info("Expired {} coupons", expiredCoupons.size());
     }
 
 }
