+++ service/coupon-api/src/main/java/com/couponify/couponapi/application/CouponIssueService.java
@@ -7,10 +7,11 @@ import com.couponify.couponapi.exception.CouponException;
 import com.couponify.coupondomain.domain.coupon.Coupon;
 import com.couponify.coupondomain.domain.coupon.CouponCache;
 import com.couponify.coupondomain.domain.coupon.repository.CouponRepository;
-import com.couponify.coupondomain.domain.issuedCoupon.repository.IssuedCouponRepository;
+import java.util.concurrent.TimeUnit;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.redisson.api.RMap;
+import org.redisson.api.RMapCache;
 import org.redisson.api.RSet;
 import org.redisson.api.RedissonClient;
 import org.springframework.stereotype.Service;
@@ -24,10 +25,8 @@ public class CouponIssueService {
 
     private static final int QUANTITY_TO_ISSUE_COUPON = 1;
     private final CouponRepository couponRepository;
-    private final IssuedCouponRepository issuedCouponRepository;
     private final RedissonClient redissonClient;
 
-    @Transactional
     public void issue(Long couponId, Long userId) {
         CouponCache couponCache = getCouponCache(couponId);
         checkUserAlreadyIssued(couponId, userId);
@@ -35,28 +34,22 @@ public class CouponIssueService {
         // TODO User 검증 필요
 
         couponCache.issue(QUANTITY_TO_ISSUE_COUPON);
-
         addIssuer(couponId, userId);
         updateCouponCache(couponId, couponCache);
     }
 
     private CouponCache getCouponCache(Long couponId) {
-        RMap<Long, CouponCache> couponInfo = redissonClient.getMap(CouponPrefix.COUPON_INFO);
+        RMapCache<Long, CouponCache> couponInfo = redissonClient.getMapCache(
+            CouponPrefix.COUPON_INFO);
         CouponCache couponCache = couponInfo.get(couponId);
 
         if (couponCache == null) {
             Coupon coupon = getCoupon(couponId);
             couponCache = CouponCache.of(coupon);
-            couponInfo.put(couponId, couponCache);
+            couponInfo.put(couponId, couponCache, 1, TimeUnit.HOURS);
         }
 
         return couponCache;
-    }
-
-    private Coupon getCoupon(Long couponId) {
-        return couponRepository.findById(couponId).orElseThrow(
-            () -> new CouponException(CouponErrorCode.COUPON_NOT_FOUND, couponId)
-        );
     }
 
     private void checkUserAlreadyIssued(Long couponId, Long userId) {
@@ -74,6 +67,12 @@ public class CouponIssueService {
     private void updateCouponCache(Long couponId, CouponCache couponCache) {
         RMap<Long, CouponCache> couponInfo = redissonClient.getMap(CouponPrefix.COUPON_INFO);
         couponInfo.put(couponId, couponCache);
+    }
+
+    private Coupon getCoupon(Long couponId) {
+        return couponRepository.findById(couponId).orElseThrow(
+            () -> new CouponException(CouponErrorCode.COUPON_NOT_FOUND, couponId)
+        );
     }
 
 }
